Combined Keyvault/Azqr into azure-compliance skill
