+++ plugin/skills/azure-compliance/references/azqr-recommendations.md
