+++ plugin/skills/azure-compliance/references/azqr-remediation-patterns.md
